// PROYECTOS/TAS-ON-WEB/tas_on_web/tas-on-web-app/src/app/interfaces/solicitudEnvio.ts
export interface SolicitudEnvio {
  solicitudEnvioId?: any;
  solicitudEnvioPeso?: any;
  solicitudEnvioUnidadPeso?: any;
  solicitudEnvioVolumen?: any;
  solicitudEnvioUnidadVolumen?: any;
  solicitudEnvioNumeroPiesas?: any;
  solicitudEnvioDireccionOrigen?: any;
  solicitudEnvioLocalidadOrigen?: any;
  solicitudEnvioDireccionDestino?: any;
  solicitudEnvioLocalidadDestino?: any;
  solicitudEnvioFechaRecoleccion?: any;
  solicitudEnvioNumeroEstibadores?: any;
  solicitudEnvioDiasValides?: any;
  solicitudEnvioFechaEntrega?: any;
  solicitudEnvioEstado?: any;
  solicitudEnvioOfertaId?: any;
  solicitudEnvioPersonaEntrega?: any;
  solicitudEnvioPersonaRecibe?: any;
  solicitudEnvioComentario?: any;
  solicitudEnvioNumeroDocCliente?: any;
  solicitudEnvioValorObjetivo?: any;
  solicitudEnvioTipoSubasta?: any;
  solicitudEnvioObservaciones?: any;
}